test: incorporate tests into the right file and dedupe
